Add multiscale bijections and supporting classes
