+++ normalizing_flows/architectures.py
@@ -25,3 +25,5 @@ from normalizing_flows.bijections.finite.residual.architectures import (
     Radial,
     Sylvester
 )
+
+from normalizing_flows.bijections.finite.multiscale.architectures import MultiscaleRealNVP
