+++ normalizing_flows/bijections/finite/autoregressive/layers_base.py
@@ -1,4 +1,4 @@
-from typing import Tuple, Union
+from typing import Tuple, Union, Type
 
 import torch
 import torch.nn as nn
